@@ -1,11 +1,11 @@
 name: wechat_assets_picker_demo
 description: The demo project for the wechat_assets_picker package.
-version: 7.2.0+16
+version: 8.0.0+20
 publish_to: none
 
 environment:
-  sdk: '>=2.15.0 <3.0.0'
-  flutter: '>=2.8.0'
+  sdk: '>=2.17.0 <3.0.0'
+  flutter: '>=3.0.0'
 
 dependencies:
   flutter:
@@ -15,10 +15,9 @@ dependencies:
 
   wechat_assets_picker:
     path: ../
-  wechat_camera_picker: ^3.0.0
+  wechat_camera_picker: ^3.2.0
 
-  meta: ^1.7.0
-  package_info_plus: ^1.3.0
+  package_info_plus: ^1.4.2
   path: ^1.8.0
   path_provider: ^2.0.8
   provider: ^6.0.2
