@@ -1,7 +1,7 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2020-05-31 21:38
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:flutter/semantics.dart';
 import 'package:flutter/services.dart';
@@ -16,10 +16,10 @@ bool get currentIsDark =>
     Screens.mediaQuery.platformBrightness == Brightness.dark;
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
-  _HomePageState createState() => _HomePageState();
+  State<HomePage> createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
@@ -108,7 +108,7 @@ class _HomePageState extends State<HomePage> {
               Expanded(
                 child: PageView(
                   controller: controller,
-                  children: <Widget>[
+                  children: const <Widget>[
                     MultiAssetsPage(),
                     SingleAssetPage(),
                     CustomPickersPage(),
