@@ -1,7 +1,7 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2021/7/13 11:46
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/widgets.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart'
     show
@@ -39,7 +39,7 @@ mixin ExamplePageMixin<T extends StatefulWidget> on State<T> {
   Future<void> selectAssets(PickMethod model) async {
     final List<AssetEntity>? result = await model.method(context, assets);
     if (result != null) {
-      assets = List<AssetEntity>.from(result);
+      assets = result.toList();
       if (mounted) {
         setState(() {});
       }
@@ -56,7 +56,7 @@ mixin ExamplePageMixin<T extends StatefulWidget> on State<T> {
 
   void onResult(List<AssetEntity>? result) {
     if (result != null && result != assets) {
-      assets = List<AssetEntity>.from(result);
+      assets = result.toList();
       if (mounted) {
         setState(() {});
       }
