@@ -1,7 +1,7 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2020-05-30 20:56
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 import 'package:wechat_camera_picker/wechat_camera_picker.dart';
@@ -92,7 +92,6 @@ class PickMethod {
           pickerConfig: AssetPickerConfig(
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
-            requestType: RequestType.common,
             specialItemPosition: SpecialItemPosition.prepend,
             specialItemBuilder: (
               BuildContext context,
@@ -140,7 +139,6 @@ class PickMethod {
           pickerConfig: AssetPickerConfig(
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
-            requestType: RequestType.common,
             specialItemPosition: SpecialItemPosition.prepend,
             specialItemBuilder: (
               BuildContext context,
@@ -166,9 +164,12 @@ class PickMethod {
                     final DefaultAssetPickerBuilderDelegate builder =
                         picker.builder as DefaultAssetPickerBuilderDelegate;
                     final DefaultAssetPickerProvider p = builder.provider;
-                    p.currentPath =
-                        await p.currentPath!.obtainForNewProperties();
-                    await p.switchPath(p.currentPath!);
+                    await p.switchPath(
+                      PathWrapper<AssetPathEntity>(
+                        path:
+                            await p.currentPath!.path.obtainForNewProperties(),
+                      ),
+                    );
                     p.selectAsset(result);
                   },
                   child: const Center(
@@ -194,7 +195,6 @@ class PickMethod {
           pickerConfig: AssetPickerConfig(
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
-            requestType: RequestType.common,
           ),
         );
       },
@@ -260,7 +260,6 @@ class PickMethod {
           pickerConfig: AssetPickerConfig(
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
-            requestType: RequestType.common,
             specialItemPosition: SpecialItemPosition.prepend,
             specialItemBuilder: (
               BuildContext context,
@@ -288,7 +287,6 @@ class PickMethod {
           pickerConfig: AssetPickerConfig(
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
-            requestType: RequestType.common,
             specialPickerType: SpecialPickerType.noPreview,
           ),
         );
@@ -306,13 +304,11 @@ class PickMethod {
       name: 'Keep scroll offset',
       description: 'Pick assets from same scroll position.',
       method: (BuildContext context, List<AssetEntity> assets) async {
-        final PermissionState _ps =
-            await PhotoManager.requestPermissionExtend();
-        if (_ps != PermissionState.authorized &&
-            _ps != PermissionState.limited) {
-          throw StateError('Permission state error with $_ps.');
+        final PermissionState ps = await PhotoManager.requestPermissionExtend();
+        if (ps != PermissionState.authorized && ps != PermissionState.limited) {
+          throw StateError('Permission state error with $ps.');
         }
-        onPermission(_ps);
+        onPermission(ps);
         return AssetPicker.pickAssetsWithDelegate(
           context,
           delegate: delegate(),
@@ -353,7 +349,7 @@ class PickMethod {
             maxAssets: maxAssetsCount,
             selectedAssets: assets,
             selectPredicate: (BuildContext c, AssetEntity a, bool isSelected) {
-              print('Asset title: ${a.title}');
+              debugPrint('Asset title: ${a.title}');
               return a.title?.endsWith('.gif') != true;
             },
           ),
