@@ -1,7 +1,7 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2020-05-31 20:21
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
@@ -9,8 +9,10 @@ import '../constants/page_mixin.dart';
 import '../constants/picker_method.dart';
 
 class MultiAssetsPage extends StatefulWidget {
+  const MultiAssetsPage({super.key});
+
   @override
-  _MultiAssetsPageState createState() => _MultiAssetsPageState();
+  State<MultiAssetsPage> createState() => _MultiAssetsPageState();
 }
 
 class _MultiAssetsPageState extends State<MultiAssetsPage>
