@@ -1,7 +1,7 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2020-05-31 21:17
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart'
     show AssetEntity;
@@ -10,8 +10,10 @@ import '../constants/page_mixin.dart';
 import '../constants/picker_method.dart';
 
 class SingleAssetPage extends StatefulWidget {
+  const SingleAssetPage({super.key});
+
   @override
-  _SingleAssetPageState createState() => _SingleAssetPageState();
+  State<SingleAssetPage> createState() => _SingleAssetPageState();
 }
 
 class _SingleAssetPageState extends State<SingleAssetPage>
