@@ -1,17 +1,17 @@
-///
-/// @Author Alex (https://github.com/AlexV525)
-/// [Date] 2021/7/13 10:40
-///
+// Copyright 2019 The FlutterCandies author. All rights reserved.
+// Use of this source code is governed by an Apache license that can be found
+// in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart'
     show AssetEntity, AssetEntityImageProvider, AssetType;
 
 class AssetWidgetBuilder extends StatelessWidget {
   const AssetWidgetBuilder({
-    Key? key,
+    super.key,
     required this.entity,
     required this.isDisplayingDetail,
-  }) : super(key: key);
+  });
 
   final AssetEntity entity;
   final bool isDisplayingDetail;
