@@ -12,6 +12,7 @@ import 'package:provider/provider.dart';
 import '../constants/constants.dart';
 import '../delegates/sort_path_delegate.dart';
 import '../internal/singleton.dart';
+import '../models/path_wrapper.dart';
 
 /// [ChangeNotifier] for assets picker.
 ///
@@ -21,13 +22,13 @@ import '../internal/singleton.dart';
 /// and how to get the thumbnail data of a path.
 abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
   AssetPickerProvider({
-    this.maxAssets = 9,
-    this.pageSize = 320,
+    this.maxAssets = defaultMaxAssetsCount,
+    this.pageSize = defaultAssetsPerPage,
     this.pathThumbnailSize = defaultPathThumbnailSize,
     List<Asset>? selectedAssets,
   }) {
-    if (selectedAssets?.isNotEmpty == true) {
-      _selectedAssets = List<Asset>.from(selectedAssets!);
+    if (selectedAssets != null && selectedAssets.isNotEmpty) {
+      _selectedAssets = selectedAssets.toList();
     }
   }
 
@@ -50,7 +51,7 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
   @override
   void dispose() {
     _isAssetsEmpty = false;
-    _pathsList.clear();
+    _paths.clear();
     _currentPath = null;
     _currentAssets.clear();
     _selectedAssets.clear();
@@ -63,11 +64,11 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
 
   /// Get the thumbnail from the first asset under the specific path entity.
   /// 获取指定路径下的第一个资源的缩略图数据
-  Future<Uint8List?> getThumbnailFromPath(Path path);
+  Future<Uint8List?> getThumbnailFromPath(PathWrapper<Path> path);
 
   /// Switch between paths.
   /// 切换路径
-  Future<void> switchPath([Path? path]);
+  Future<void> switchPath([PathWrapper<Path>? path]);
 
   /// Get assets under the specific path entity.
   /// 获取指定路径下的资源
@@ -105,7 +106,7 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
 
   /// Whether more assets are waiting for a load.
   /// 是否还有更多资源可以加载
-  bool get hasMoreToLoad => _currentAssets.length < _totalAssetsCount;
+  bool get hasMoreToLoad => _currentAssets.length < _totalAssetsCount!;
 
   /// The current page for assets list.
   /// 当前加载的资源列表分页数
@@ -114,10 +115,10 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
 
   /// Total count for assets.
   /// 资源总数
-  int get totalAssetsCount => _totalAssetsCount;
-  int _totalAssetsCount = 0;
+  int? get totalAssetsCount => _totalAssetsCount;
+  int? _totalAssetsCount;
 
-  set totalAssetsCount(int value) {
+  set totalAssetsCount(int? value) {
     if (value == _totalAssetsCount) {
       return;
     }
@@ -131,31 +132,30 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
   /// Using [Map] in order to save the thumbnail data
   /// for the first asset under the path.
   /// 使用 [Map] 来保存路径下第一个资源的缩略图数据
-  Map<Path, Uint8List?> get pathsList => _pathsList;
-  final Map<Path, Uint8List?> _pathsList = <Path, Uint8List?>{};
+  List<PathWrapper<Path>> get paths => _paths;
+  final List<PathWrapper<Path>> _paths = <PathWrapper<Path>>[];
 
   /// Set thumbnail [data] for the specific [path].
   /// 为指定的路径设置缩略图数据
   void setPathThumbnail(Path path, Uint8List? data) {
-    _pathsList[path] = data;
-    notifyListeners();
+    final int index = _paths.indexWhere(
+      (PathWrapper<Path> w) => w.path == path,
+    );
+    if (index != -1) {
+      final PathWrapper<Path> newWrapper = _paths[index].copyWith(
+        thumbnailData: data,
+      );
+      _paths[index] = newWrapper;
+      notifyListeners();
+    }
   }
-
-  /// How many path has a valid thumb data.
-  /// 当前有多少目录已经正常载入了缩略图
-  ///
-  /// This getter provides a "Should Rebuild" condition judgement to [Selector]
-  /// with the path entities widget.
-  /// 它为目录部件展示部分的 [Selector] 提供了是否重建的条件。
-  int get validPathThumbnailsCount =>
-      _pathsList.values.where((Uint8List? d) => d != null).length;
 
   /// The path which is currently using.
   /// 正在查看的资源路径
-  Path? get currentPath => _currentPath;
-  Path? _currentPath;
+  PathWrapper<Path>? get currentPath => _currentPath;
+  PathWrapper<Path>? _currentPath;
 
-  set currentPath(Path? value) {
+  set currentPath(PathWrapper<Path>? value) {
     if (value == _currentPath) {
       return;
     }
@@ -172,7 +172,7 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
     if (value == _currentAssets) {
       return;
     }
-    _currentAssets = List<Asset>.from(value);
+    _currentAssets = value.toList();
     notifyListeners();
   }
 
@@ -185,7 +185,7 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
     if (value == _selectedAssets) {
       return;
     }
-    _selectedAssets = List<Asset>.from(value);
+    _selectedAssets = value.toList();
     notifyListeners();
   }
 
@@ -212,39 +212,36 @@ abstract class AssetPickerProvider<Asset, Path> extends ChangeNotifier {
     if (selectedAssets.length == maxAssets || selectedAssets.contains(item)) {
       return;
     }
-    final List<Asset> _set = List<Asset>.from(selectedAssets);
-    _set.add(item);
-    selectedAssets = _set;
+    final List<Asset> set = selectedAssets.toList();
+    set.add(item);
+    selectedAssets = set;
   }
 
   /// Un-select asset.
   /// 取消选中资源
   void unSelectAsset(Asset item) {
-    final List<Asset> _set = List<Asset>.from(selectedAssets);
-    _set.remove(item);
-    selectedAssets = _set;
+    final List<Asset> set = selectedAssets.toList();
+    set.remove(item);
+    selectedAssets = set;
   }
 }
 
 class DefaultAssetPickerProvider
     extends AssetPickerProvider<AssetEntity, AssetPathEntity> {
   DefaultAssetPickerProvider({
-    List<AssetEntity>? selectedAssets,
+    super.selectedAssets,
+    super.maxAssets,
+    super.pageSize,
+    super.pathThumbnailSize,
     this.requestType = RequestType.image,
     this.sortPathDelegate = SortPathDelegate.common,
+    this.sortPathsByModifiedDate = false,
     this.filterOptions,
-    int maxAssets = 9,
-    int pageSize = 80,
-    ThumbnailSize pathThumbnailSize = const ThumbnailSize.square(80),
-  }) : super(
-          maxAssets: maxAssets,
-          pageSize: pageSize,
-          pathThumbnailSize: pathThumbnailSize,
-          selectedAssets: selectedAssets,
-        ) {
+    Duration initializeDelayDuration = const Duration(milliseconds: 250),
+  }) {
     Singleton.sortPathDelegate = sortPathDelegate ?? SortPathDelegate.common;
     // Call [getAssetList] with route duration when constructing.
-    Future<void>(() async {
+    Future<void>.delayed(initializeDelayDuration, () async {
       await getPaths();
       await getAssetsFromCurrentPath();
     });
@@ -252,19 +249,15 @@ class DefaultAssetPickerProvider
 
   @visibleForTesting
   DefaultAssetPickerProvider.forTest({
-    List<AssetEntity>? selectedAssets,
+    super.selectedAssets,
     this.requestType = RequestType.image,
     this.sortPathDelegate = SortPathDelegate.common,
+    this.sortPathsByModifiedDate = false,
     this.filterOptions,
-    int maxAssets = 9,
-    int pageSize = 80,
-    ThumbnailSize pathThumbnailSize = const ThumbnailSize.square(80),
-  }) : super(
-          maxAssets: maxAssets,
-          pageSize: pageSize,
-          pathThumbnailSize: pathThumbnailSize,
-          selectedAssets: selectedAssets,
-        ) {
+    super.maxAssets,
+    super.pageSize = 80,
+    super.pathThumbnailSize,
+  }) {
     Singleton.sortPathDelegate = sortPathDelegate ?? SortPathDelegate.common;
   }
 
@@ -272,9 +265,11 @@ class DefaultAssetPickerProvider
   /// 请求的资源类型
   final RequestType requestType;
 
-  /// Delegate to sort asset path entities.
-  /// 资源路径排序的实现
+  /// @{macro wechat_assets_picker.delegates.SortPathDelegate}
   final SortPathDelegate<AssetPathEntity>? sortPathDelegate;
+
+  /// {@macro wechat_assets_picker.constants.AssetPickerConfig.sortPathsByModifiedDate}
+  final bool sortPathsByModifiedDate;
 
   /// Filter options for the picker.
   /// 选择器的筛选条件
@@ -284,25 +279,19 @@ class DefaultAssetPickerProvider
   final FilterOptionGroup? filterOptions;
 
   @override
-  set currentPath(AssetPathEntity? value) {
+  set currentPath(PathWrapper<AssetPathEntity>? value) {
     if (value == _currentPath) {
       return;
     }
     _currentPath = value;
-    if (value != null &&
-        _pathsList.keys.any((AssetPathEntity p) => p.id == value.id)) {
-      final AssetPathEntity previous = _pathsList.keys.singleWhere(
-        (AssetPathEntity p) => p.id == value.id,
+    if (value != null) {
+      final int index = _paths.indexWhere(
+        (PathWrapper<AssetPathEntity> p) => p.path.id == value.path.id,
       );
-      final int index = _pathsList.keys.toList().indexOf(previous);
-      final List<MapEntry<AssetPathEntity, Uint8List?>> newEntries =
-          _pathsList.entries.toList()
-            ..removeAt(index)
-            ..insert(index, MapEntry<AssetPathEntity, Uint8List?>(value, null));
-      _pathsList
-        ..clear()
-        ..addEntries(newEntries);
-      getThumbnailFromPath(value);
+      if (index != -1) {
+        _paths[index] = value;
+        getThumbnailFromPath(value);
+      }
     }
     notifyListeners();
   }
@@ -320,7 +309,7 @@ class DefaultAssetPickerProvider
         needTitle: true,
         sizeConstraint: SizeConstraint(ignoreSize: true),
       ),
-      containsPathModified: true,
+      containsPathModified: sortPathsByModifiedDate,
     );
 
     // Merge user's filter option into base options if it's not null.
@@ -328,104 +317,97 @@ class DefaultAssetPickerProvider
       options.merge(filterOptions!);
     }
 
-    final List<AssetPathEntity> _list = await PhotoManager.getAssetPathList(
+    final List<AssetPathEntity> list = await PhotoManager.getAssetPathList(
       type: requestType,
       filterOption: options,
     );
 
-    // Sort path using sort path delegate.
-    Singleton.sortPathDelegate.sort(_list);
-
-    for (final AssetPathEntity pathEntity in _list) {
-      // Use sync method to avoid unnecessary wait.
-      _pathsList[pathEntity] = null;
-      getThumbnailFromPath(pathEntity);
+    for (final AssetPathEntity pathEntity in list) {
+      final int index = _paths.indexWhere(
+        (PathWrapper<AssetPathEntity> p) => p.path.id == pathEntity.id,
+      );
+      final PathWrapper<AssetPathEntity> wrapper = PathWrapper<AssetPathEntity>(
+        path: pathEntity,
+      );
+      if (index == -1) {
+        _paths.add(wrapper);
+      } else {
+        _paths[index] = wrapper;
+      }
     }
+    // Sort path using sort path delegate.
+    Singleton.sortPathDelegate.sort(_paths);
+    // Use sync method to avoid unnecessary wait.
+    _paths
+      ..forEach(getAssetCountFromPath)
+      ..forEach(getThumbnailFromPath);
 
     // Set first path entity as current path entity.
-    if (_pathsList.isNotEmpty) {
-      _currentPath ??= pathsList.keys.elementAt(0);
+    if (_paths.isNotEmpty) {
+      _currentPath ??= _paths.first;
     }
   }
 
   @override
-  Future<void> getAssetsFromPath(int page, AssetPathEntity path) async {
+  Future<void> getAssetsFromPath([int? page, AssetPathEntity? path]) async {
+    page ??= currentAssetsListPage;
+    path ??= currentPath!.path;
     final List<AssetEntity> list = await path.getAssetListPaged(
       page: page,
       size: pageSize,
     );
-    _currentAssets = List<AssetEntity>.of(list);
+    if (page == 0) {
+      _currentAssets = list;
+    } else {
+      _currentAssets.addAll(list);
+    }
     _hasAssetsToDisplay = currentAssets.isNotEmpty;
     notifyListeners();
   }
 
   @override
-  Future<void> loadMoreAssets() async {
-    final List<AssetEntity> list = await currentPath!.getAssetListPaged(
-      page: currentAssetsListPage,
-      size: pageSize,
-    );
-    final List<AssetEntity> assets = List<AssetEntity>.of(list);
-    if (assets.isNotEmpty && currentAssets.contains(assets[0])) {
-      return;
-    }
-    final List<AssetEntity> tempList = <AssetEntity>[];
-    tempList.addAll(_currentAssets);
-    tempList.addAll(assets);
-    currentAssets = tempList;
-  }
+  Future<void> loadMoreAssets() => getAssetsFromPath();
 
   @override
-  Future<void> switchPath([AssetPathEntity? path]) async {
+  Future<void> switchPath([PathWrapper<AssetPathEntity>? path]) async {
     assert(
       () {
-        if (_currentPath == null && path == null) {
+        if (path == null && _currentPath == null) {
           throw FlutterError.fromParts(<DiagnosticsNode>[
-            ErrorSummary('Empty $AssetPathEntity was switched.'),
+            ErrorSummary('Switching empty path.'),
             ErrorDescription(
-              'Neither currentPathEntity nor pathEntity is non-null, '
+              'Neither "path" nor "currentPathEntity" is non-null, '
               'which makes this method useless.',
             ),
             ErrorHint(
-              'You need to pass a non-null $AssetPathEntity '
-              'or call this method when currentPathEntity is not null.',
+              'You need to pass a non-null path or call this method '
+              'when the "currentPath" is not null.',
             ),
           ]);
         }
         return true;
       }(),
     );
-    if (_currentPath == null && path == null) {
+    if (path == null && _currentPath == null) {
       return;
     }
     path ??= _currentPath!;
     _currentPath = path;
-    _totalAssetsCount = path.assetCount;
-    notifyListeners();
-    await getAssetsFromPath(0, currentPath!);
+    await getAssetsFromCurrentPath();
   }
 
   @override
   Future<Uint8List?> getThumbnailFromPath(
-    AssetPathEntity path,
+    PathWrapper<AssetPathEntity> path,
   ) async {
-    assert(
-      () {
-        if (path.assetCount < 1) {
-          throw FlutterError.fromParts(<DiagnosticsNode>[
-            ErrorSummary('No assets in the path ${path.id}.'),
-            ErrorDescription(
-              'Thumbnail can only obtained when the path contains assets.',
-            ),
-          ]);
-        }
-        return true;
-      }(),
-    );
     if (requestType == RequestType.audio) {
       return null;
     }
-    final List<AssetEntity> assets = await path.getAssetListRange(
+    final int assetCount = path.assetCount ?? await path.path.assetCountAsync;
+    if (assetCount == 0) {
+      return null;
+    }
+    final List<AssetEntity> assets = await path.path.getAssetListRange(
       start: 0,
       end: 1,
     );
@@ -437,20 +419,45 @@ class DefaultAssetPickerProvider
     if (asset.type != AssetType.image && asset.type != AssetType.video) {
       return null;
     }
-    final Uint8List? assetData = await asset.thumbnailDataWithSize(
+    final Uint8List? data = await asset.thumbnailDataWithSize(
       pathThumbnailSize,
     );
-    _pathsList[path] = assetData;
-    notifyListeners();
-    return assetData;
+    final int index = _paths.indexWhere(
+      (PathWrapper<AssetPathEntity> p) => p.path == path.path,
+    );
+    if (index != -1) {
+      _paths[index] = _paths[index].copyWith(thumbnailData: data);
+      notifyListeners();
+    }
+    return data;
+  }
+
+  Future<void> getAssetCountFromPath(PathWrapper<AssetPathEntity> path) async {
+    final int assetCount = await path.path.assetCountAsync;
+    final int index = _paths.indexWhere(
+      (PathWrapper<AssetPathEntity> p) => p == path,
+    );
+    if (index != -1) {
+      _paths[index] = _paths[index].copyWith(assetCount: assetCount);
+      if (index == 0) {
+        _currentPath = _currentPath?.copyWith(assetCount: assetCount);
+      }
+      notifyListeners();
+    }
   }
 
   /// Get assets list from current path entity.
   /// 从当前已选路径获取资源列表
   Future<void> getAssetsFromCurrentPath() async {
-    if (_pathsList.isNotEmpty) {
-      totalAssetsCount = currentPath!.assetCount;
-      await getAssetsFromPath(0, currentPath!);
+    if (_currentPath != null && _paths.isNotEmpty) {
+      final PathWrapper<AssetPathEntity> wrapper = _currentPath!;
+      final int assetCount =
+          wrapper.assetCount ?? await wrapper.path.assetCountAsync;
+      totalAssetsCount = assetCount;
+      if (wrapper.assetCount == null) {
+        currentPath = _currentPath!.copyWith(assetCount: assetCount);
+      }
+      await getAssetsFromPath(0, currentPath!.path);
     } else {
       isAssetsEmpty = true;
     }
