@@ -2,16 +2,25 @@
 // Use of this source code is governed by an Apache license that can be found
 // in the LICENSE file.
 
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+import '../constants/constants.dart';
 
 /// [ChangeNotifier] for assets picker viewer.
 /// 资源选择查看器的 provider model.
 class AssetPickerViewerProvider<A> extends ChangeNotifier {
   /// Copy selected assets for editing when constructing.
   /// 构造时深拷贝已选择的资源集合，用于后续编辑。
-  AssetPickerViewerProvider(List<A>? assets) {
-    _currentlySelectedAssets = List<A>.from(assets ?? <A>[]);
+  AssetPickerViewerProvider(
+    List<A>? assets, {
+    this.maxAssets = defaultMaxAssetsCount,
+  }) {
+    _currentlySelectedAssets = (assets ?? <A>[]).toList();
   }
+
+  /// Maximum count for asset selection.
+  /// 资源选择的最大数量
+  final int maxAssets;
 
   /// Selected assets in the viewer.
   /// 查看器中已选择的资源
@@ -32,17 +41,29 @@ class AssetPickerViewerProvider<A> extends ChangeNotifier {
 
   /// Select asset.
   /// 选中资源
-  void selectAssetEntity(A entity) {
-    final List<A> set = List<A>.from(currentlySelectedAssets);
-    set.add(entity);
-    currentlySelectedAssets = List<A>.from(set);
+  void selectAsset(A item) {
+    if (currentlySelectedAssets.length == maxAssets ||
+        currentlySelectedAssets.contains(item)) {
+      return;
+    }
+    final List<A> newList = _currentlySelectedAssets.toList()..add(item);
+    currentlySelectedAssets = newList;
   }
 
   /// Un-select asset.
   /// 取消选中资源
-  void unSelectAssetEntity(A entity) {
-    final List<A> set = List<A>.from(currentlySelectedAssets);
-    set.remove(entity);
-    currentlySelectedAssets = List<A>.from(set);
+  void unSelectAsset(A item) {
+    if (currentlySelectedAssets.isEmpty ||
+        !currentlySelectedAssets.contains(item)) {
+      return;
+    }
+    final List<A> newList = _currentlySelectedAssets.toList()..remove(item);
+    currentlySelectedAssets = newList;
   }
+
+  @Deprecated('Use selectAsset instead')
+  void selectAssetEntity(A entity) => selectAsset(entity);
+
+  @Deprecated('Use unSelectAsset instead')
+  void unselectAssetEntity(A entity) => unSelectAsset(entity);
 }
