@@ -13,18 +13,18 @@ import '../scale_text.dart';
 
 class LocallyAvailableBuilder extends StatefulWidget {
   const LocallyAvailableBuilder({
-    Key? key,
+    super.key,
     required this.asset,
     required this.builder,
     this.isOriginal = true,
-  }) : super(key: key);
+  });
 
   final AssetEntity asset;
   final Widget Function(BuildContext context, AssetEntity asset) builder;
   final bool isOriginal;
 
   @override
-  _LocallyAvailableBuilderState createState() =>
+  State<LocallyAvailableBuilder> createState() =>
       _LocallyAvailableBuilderState();
 }
 
