@@ -7,6 +7,7 @@ import 'package:flutter/material.dart';
 class ScaleText extends StatelessWidget {
   const ScaleText(
     this.text, {
+    super.key,
     this.style,
     this.strutStyle,
     this.maxLines,
@@ -33,8 +34,10 @@ class ScaleText extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final MediaQueryData mqd = MediaQuery.of(context);
-    final double effectiveFactor =
-        mqd.textScaleFactor.clamp(minScaleFactor, maxScaleFactor).toDouble();
+    final double effectiveFactor = mqd.textScaleFactor.clamp(
+      minScaleFactor,
+      maxScaleFactor,
+    );
     return MediaQuery(
       data: mqd.copyWith(textScaleFactor: effectiveFactor),
       child: Text(
